requirements.txt
fastapi==0.38.1
tortoise-orm==0.13.5
uvicorn==0.9.0